+++ profiles/ulf/themes/ulf_default/templates/fields/field--field-contact-name.tpl.php
@@ -1,0 +1,3 @@
+<?php foreach ($items as $delta => $item): ?>
+  <div class="field--collection-description is-bold"><?php print render($item); ?></div>
+<?php endforeach; ?>