+++ profiles/ulf/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #v2.2.0
 * Updated install profile to reflect ULF version 2.0.0
+* Added Silkeborg theme
 
 #v2.1.2